remove limsSample class, add LimsMapper that creates SampleUpdate objects
